move openid class to openid::dynamic
