+++ lib/rubygems/sigstore/cert_provider.rb
@@ -5,7 +5,7 @@ class Gem::Sigstore::CertProvider
   end
 
   def run
-    proof, access_token = Gem::Sigstore::OpenID.new(pkey.private_key).get_token
+    proof, access_token = Gem::Sigstore::OpenID::Dynamic.new(pkey.private_key).get_token
     fulcio_api = Gem::Sigstore::FulcioApi.new(token: access_token, host: config.fulcio_host)
     fulcio_api.create(proof, pkey.public_key.to_der)
   end
